requirements.txt
langchain
pypdf2
python-dotenv
streamlit
openai
